Fixed bug with invalid id assignation on alarm creation. Alarm id type migrated to long

+++ app/src/main/java/sleepless_nights/location_alarm/geofence/CustomGeofence.java
@@ -17,7 +17,7 @@ import sleepless_nights.location_alarm.geofence.use_cases.GeofenceBroadcastRecei
 public class CustomGeofence {
     public static final int GEOFENCE_TRANSITION_TYPES = Geofence.GEOFENCE_TRANSITION_ENTER;
 
-    private int alarmId;
+    private long alarmId;
     private Geofence geofence;
     private PendingIntent geofencePendingIntent;
     private Context context;
@@ -61,5 +61,5 @@ public class CustomGeofence {
     @NonNull
     public String getGeofenceId() { return geofence.getRequestId(); }
 
-    public int getAlarmId() { return alarmId; }
+    public long getAlarmId() { return alarmId; }
 }
